chore: cambiar campo cantidad de INT a DECIMAL(6,1) en ordenes_productos
Permite cantidades fraccionarias como 2.5, 0.5, etc.

+++ public/model/create_new_company_api_emp_N.sql
@@ -600,7 +600,7 @@ CREATE TABLE `ordenes_productos` (
   `id_products_attributes` int(11) DEFAULT NULL COMMENT 'ID de la variante del producto',
   `category_name` varchar(20) DEFAULT NULL COMMENT 'NOMBRE de la categoria en woocommerce',
   `name` varchar(240) DEFAULT NULL COMMENT 'Nombre del producto',
-  `cantidad` int(11) NOT NULL DEFAULT 0 COMMENT 'Cantidad del producto',
+  `cantidad` DECIMAL(6,1) NOT NULL DEFAULT 0 COMMENT 'Cantidad del producto',
   `id_size` int(11) DEFAULT NULL COMMENT 'ID de la talla',
   `talla` varchar(8) DEFAULT NULL COMMENT 'Talla del producto',
   `corte` varchar(32) DEFAULT NULL COMMENT 'Dama, caballero, niño',
